requirements.txt
opencv-python==4.8.1.78
face-recognition==1.3.5
mysql-connector-python==8.2.0
pandas==2.1.3
python-dotenv==1.0.0
